backend/pyproject.toml
[tool.poetry]
name = "sentiment-labeller"
version = "0.1.0"
description = ""
authors = ["HasiNed <hasined@foxmail.com>"]
packages = [{ include = "app" }]

[tool.poetry.scripts]
dev = "run:dev"
start = "run:start"
lint = "run:lint"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.0"
ruff = "^0.0.226"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
